requirements.txt
Django==4.1.3
djangorestframework==3.14.0
apscheduler==3.9.1
requests==2.28.1
drf-spectacular==0.24.2